revert the handler
