+++ test/test_helper.exs
@@ -21,8 +21,6 @@ Mox.defmock(MockMetricsCollector, for: Archethic.Metrics.Collector)
 # -----Start-of-Networking-Mocks-----
 
 Mox.defmock(MockStatic, for: Archethic.Networking.IPLookup.Impl)
-Mox.defmock(MockLocalHandler, for: Archethic.Networking.IPLookup.Impl)
-Mox.defmock(MockRemoteHandler, for: Archethic.Networking.IPLookup.Impl)
 Mox.defmock(MockIPLookup, for: Archethic.Networking.IPLookup.Impl)
 
 # -----End-of-Networking-Mocks ------
